Refactor: Remove old test classes and implement new test structure
- Deleted existing test classes: AtlasGameHistoryTest, AtlasGameTest, AtlasPlaceValidatorProviderTest, AtlasPlaceValidatorTest, and TestRunner.
- Created new test classes with updated implementations for AtlasGameHistory, AtlasGame, AtlasPlaceValidatorProvider, and AtlasPlaceValidator.
- Updated TestRunner to reflect new test structure and ensure all tests are executed.
- Added new exception handling in AtlasExceptions for better error management.
- Introduced AtlasGame and AtlasGameHistory classes with enhanced functionality.
- Implemented AtlasPlaceValidator with file reading capabilities for valid places.
- Added a resource file (ValidPlaces.txt) for place validation.
- Updated build.gradle to include necessary configurations for running tests.
