+++ src/main/java/com/AtlasLibrary/AtlasGameHistory.java
@@ -1,4 +1,5 @@
-package main.java;
+package com.AtlasLibrary;
+
 import java.util.ArrayList;
 import java.util.List;
