+++ src/main/java/com/AtlasLibrary/AtlasPlaceValidatorProvider.java
@@ -1,4 +1,5 @@
-package main.java;
+package com.AtlasLibrary;
+
 public class AtlasPlaceValidatorProvider {
     private static final AtlasPlaceValidator atlasPlaceValidator = new AtlasPlaceValidator();
     private AtlasPlaceValidatorProvider()
