+++ src/main/resources/ValidPlaces.txt
