+++ src/test/java/AtlasGameHistoryTest.java
@@ -1,7 +1,4 @@
-package test.java;
-import main.java.AtlasGameHistory;
-import static test.java.TestRunner.AssertEquals;
-
+import com.AtlasLibrary.AtlasGameHistory;
 
 class AtlasGameHistoryTest {
 
@@ -13,8 +10,8 @@ class AtlasGameHistoryTest {
         int player=1;
         String word="Khammam";
         atlasGameHistory.recordMove(player, word);
-        AssertEquals(atlasGameHistory.getPlayerOnTurn(0),player);
-        AssertEquals(atlasGameHistory.getActionOnTurn(0),word);
+        TestRunner.AssertEquals(atlasGameHistory.getPlayerOnTurn(0),player);
+        TestRunner.AssertEquals(atlasGameHistory.getActionOnTurn(0),word);
         return true;
     }
 }
