+++ src/test/java/AtlasPlaceValidatorProviderTest.java
@@ -1,7 +1,7 @@
-package test.java;
-import main.java.AtlasPlaceValidator;
-import main.java.AtlasPlaceValidatorProvider;
-import static test.java.TestRunner.AssertEquals;
+
+import com.AtlasLibrary.AtlasPlaceValidator;
+import com.AtlasLibrary.AtlasPlaceValidatorProvider;
+
 
 
 class AtlasPlaceValidatorProviderTest {
@@ -12,7 +12,7 @@ class AtlasPlaceValidatorProviderTest {
     public static boolean singletonProvisionTest() throws TestRunner.AssertBooleansException {
         AtlasPlaceValidator a = AtlasPlaceValidatorProvider.getAtlasPlaceValidator();
         AtlasPlaceValidator b = AtlasPlaceValidatorProvider.getAtlasPlaceValidator();
-        AssertEquals(a==b, Boolean.TRUE);
+        TestRunner.AssertEquals(a==b, Boolean.TRUE);
         return true;
     }
 }
